requirements.txt
alembic==1.15.2
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
PyJWT==2.10.1
python-docx==1.1.2
python-dotenv==1.1.0
RapidFuzz==3.13.0
requests==2.32.3
sentry-sdk==2.29.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
